Worked marine accesories foresale

+++ app/modules/scrap/controllers/ScrapController.php
@@ -43,7 +43,7 @@ class ScrapController extends \BaseController {
     public function index() {
 
 
-        $durations = Forsale::all();
+        $durations = Scrap::all();
 
 //
 //        $items = array();
@@ -80,6 +80,7 @@ class ScrapController extends \BaseController {
         $rules = array(
             'uuid'      => 'required',
             'make_region'       => 'required',
+            'model'       => 'required',
             'title'       => 'required',
             'phone'       => 'required',
             'price'       => 'required',
@@ -154,6 +155,7 @@ class ScrapController extends \BaseController {
                                 $forsale = new Scrap();
                                 $forsale->make_region_id        = $makeRegion->make_region_id;
                                 $forsale->user_scrap_id        = $mobileUser->user_scrap_id;
+                                $forsale->model        = trim(Input::get('model'));
                                 $forsale->title        = trim(Input::get('title'));
                                 $forsale->phone         = trim(Input::get('phone'));
                                 $forsale->price         = trim(Input::get('price'));
@@ -387,6 +389,7 @@ class ScrapController extends \BaseController {
                         // Add forsale data
                         $forsale = new Scrap();
                         $forsale->make_region_id        = $makeRegion->make_region_id;
+                        $forsale->model        = trim(Input::get('model'));
                         $forsale->title        = trim(Input::get('title'));
                         $forsale->phone         = trim(Input::get('phone'));
                         $forsale->price         = trim(Input::get('price'));
@@ -613,48 +616,223 @@ class ScrapController extends \BaseController {
 //    }
 
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  int  $id
-     * @return Response
-     * curl -i -X PUT --user admin:admin -d 'title=Updated Title' localhost/project/api/v1/pages/2
-     */
 
-    public function update($id) {
+    public function update() {
 
-        $input = Input::all();
 
-        $page = Duration::find($id);
-
-        if ( !empty($input['level']) ) {
-            $page->level =$input['level'];
-        }
-
-        if ( !empty($input['date']) ) {
-            $page->date =$input['date'];
-        }
-
-        if ( !empty($input['time']) ) {
-            $page->time =$input['time'];
-        }
-
-        if ( !empty($input['location']) ) {
-            $page->location =$input['location'];
-        }
-
-        if ( !empty($input['status']) ) {
-            $page->status =$input['status'];
-        }
-
-        $page->save();
-
-        return Response::json(array(
-            'error' => false,
-            'message' => 'Page Updated'),
-            200
+        $response = array();
+        // validate
+        $rules = array(
+            'id'      => 'required',
+            'uuid'      => 'required',
+            //'model'       => 'required',
+            //'title'       => 'required',
+            //'phone'       => 'required',
+            //'price'       => 'required',
+            //'description'   => 'required',
         );
+
+        $validator = Validator::make(Input::all(), $rules);
+
+        // process the login
+        if ($validator->fails()) {
+            $response = array( 'status'=> 'fail', 'message'=> $validator->messages()->first() );
+        } else {
+
+            $mobileUser = Userscrap::where('uuid','=',trim(Input::get('uuid')))->first();
+
+            //User already registered
+            if( !empty($mobileUser->user_scrap_id) && $mobileUser->status == '1'){
+
+                $id = trim(Input::get('id'));
+                $forsale = Scrap::find($id);
+
+                if( !empty($forsale->scrap_id) && $forsale->user_scrap_id == $mobileUser->user_scrap_id ){
+
+                    //$forsale->make_id        = $make->make_id;
+                    //$forsale->user_mobile_id        = $mobileUser->user_mobile_id;
+                    $model = trim(Input::get('model'));
+                    if( !empty($model) ){
+                        $forsale->model        = $model;
+                    }
+
+                    $title = trim(Input::get('title'));
+                    if( !empty($title) ){
+                        $forsale->title        = $title;
+                    }
+
+                    $phone = trim(Input::get('phone'));
+                    if( !empty($phone) ){
+                        $forsale->phone        = $phone;
+                    }
+
+                    $price = trim(Input::get('price'));
+                    if( !empty($price) ){
+                        $forsale->price        = $price;
+                    }
+
+                    $description = trim(Input::get('description'));
+                    if( !empty($description) ){
+                        $forsale->description        = $description;
+                    }
+
+
+                    //$forsale->save();
+
+                    if( ! $forsale->save() ){
+                        $response = array( 'status'=> 'fail', 'message'=> 'Something went wrong.' );
+                    }else{
+
+                        //Update quota for this user
+
+                        //Upload Images, new logic - Start here
+                        $image_sizes =  \Config::get('constant.image_sizes');
+
+                        $imagePath = public_path('uploads') .'/images/';
+
+                        $filnalDestinationPath = $imagePath . "scrap/";
+
+                        if( !file_exists($filnalDestinationPath) ){
+                            mkdir($filnalDestinationPath, 0755, true);
+                        }
+
+                        //Start - Delete old images
+                        $photos = Photos::where('scrap_id', '=', $forsale->scrap_id)->get();
+                        foreach($photos as $photo){
+
+                            $photo_name = $photo->photo_name;
+
+                            //Upload original image
+                            $fileOriginalPath = $filnalDestinationPath . "original/" . $photo_name;
+
+                            if( !empty($image_sizes) && is_array($image_sizes) ){
+
+                                foreach($image_sizes as $image_size){
+
+                                    $filePathImageSize = $filnalDestinationPath . "$image_size/" . $photo_name;
+
+                                    //Delete file
+                                    unlink($filePathImageSize);
+
+                                    //Unset
+                                    unset($filePathImageSize);
+
+
+                                }
+
+
+                            }
+
+
+                            //delete original image
+                            unlink($fileOriginalPath);
+
+                            //Delete photo
+                            $photo->delete();
+
+                            unset($photo_name);
+
+
+
+                        }
+
+                        unset($photos,$photo,$fileOriginalPath);
+                        //End - old image deleted
+
+
+
+                        $imagenumber = \Config::get('constant.number_of_image',10);
+                        for($i=1; $i <= $imagenumber; $i++){
+
+                            if (Input::hasFile('image'.$i) ) {
+                                $file            = Input::file('image'.$i);
+                                $validrules = array('file' => 'required|mimes:png,gif,jpeg,jpg'); // 'required|mimes:png,gif,jpeg,txt,pdf,doc'
+                                $validator = Validator::make(array('file'=> $file), $validrules);
+
+                                if($validator->passes()) {
+
+
+                                    //get File Name
+                                    $fileExtension = $file->getClientOriginalExtension();
+
+                                    if( empty($fileExtension) ){
+                                        continue;
+                                    }
+
+                                    $newFilename = uniqid(). "_" . time() . '.' . $fileExtension;
+
+                                    //Upload original image
+                                    $fileOriginalPath = $filnalDestinationPath . "original/" . $newFilename;
+                                    \Image::make( $file->getRealPath() )
+                                        ->save($fileOriginalPath);
+
+
+
+                                    if( !empty($image_sizes) && is_array($image_sizes) ){
+
+                                        foreach($image_sizes as $image_size){
+
+                                            $filePathImageSize = $filnalDestinationPath . "$image_size/" . $newFilename;
+
+                                            \Image::make( $file->getRealPath() )
+                                                ->resize($image_size,null, function ($constraint) { $constraint->aspectRatio(); })
+                                                ->save($filePathImageSize);
+
+                                            //Unset
+                                            unset($filePathImageSize);
+                                            unset($image_size);
+
+                                        }
+
+
+                                    }
+
+                                    //Save into db
+                                    $photo = new Photos();
+                                    $photo->scrap_id       = $forsale->scrap_id;
+                                    $photo->photo_name       = $newFilename;
+                                    $photo->save();
+
+                                }else{
+                                    continue;
+                                }
+
+                                //Unset
+                                unset($fileExtension);
+                                unset($newFilename);
+                                unset($photo);
+
+                            }else{
+                                continue;
+                            }
+
+
+                        }
+
+                        $response = array( 'status'=> 'success', 'message'=> 'Successfully Updated.' );
+
+                    }
+
+                }else{
+                    $response = array( 'status'=> 'fail', 'message'=> 'Unauthorize access.' );
+                }
+
+
+            }elseif( !empty($mobileUser->user_scrap_id) && empty($mobileUser->status) ){
+
+                $response = array( 'status'=> 'fail', 'message'=> 'Unauthorize access.' );
+
+            }else{
+                //New User
+                $response = array( 'status'=> 'fail', 'message'=> 'Something went wrong.' );
+            }
+        }
+
+        return $response ;
+
     }
+
+
 
     /**
      * Remove the specified resource from storage.
@@ -664,17 +842,103 @@ class ScrapController extends \BaseController {
      * curl -i -X DELETE --user admin:admin localhost/project/api/v1/pages/1
      */
 
-    public function destroy($id) {
-        $page = Duration::find($id);
 
-        $page->delete();
+    public function destroy() {
 
-        return Response::json(array(
-            'error' => false,
-            'message' => 'Page Deleted'),
-            200
+        $rules = array(
+            'device_phone'      => 'required',
+            'delete_id'       => 'required'
         );
+//echo '<pre>';print_r(Input::hasFile('images*'));die('======Debugging=======');
+        $validator = Validator::make(Input::all(), $rules);
+
+        // process the login
+        if ($validator->fails()) {
+            $response = array( 'status'=> 'fail', 'message'=> $validator->messages()->first() );
+        } else {
+
+            $mobileUser = Userscrap::where('uuid', '=', trim(Input::get('device_phone')))->first();
+
+            //User already registered
+            if (!empty($mobileUser->user_scrap_id) && $mobileUser->status == '1') {
+
+                $delete_id = trim(Input::get('delete_id'));
+                $page = Scrap::find($delete_id);
+
+                if( !empty($page->scrap_id) &&  $page->user_scrap_id == $mobileUser->user_scrap_id ){
+
+
+                    //Upload Images, new logic - Start here
+                    $image_sizes =  \Config::get('constant.image_sizes');
+
+                    $imagePath = public_path('uploads') .'/images/';
+
+                    $filnalDestinationPath = $imagePath . "scrap/";
+
+
+                    //Delete Photos
+                    $photos = Photos::where('scrap_id', '=', $page->scrap_id)->get();
+                    foreach($photos as $photo){
+
+                        $photo_name = $photo->photo_name;
+
+                        //Upload original image
+                        $fileOriginalPath = $filnalDestinationPath . "original/" . $photo_name;
+
+
+
+
+                        if( !empty($image_sizes) && is_array($image_sizes) ){
+
+                            foreach($image_sizes as $image_size){
+
+                                $filePathImageSize = $filnalDestinationPath . "$image_size/" . $photo_name;
+
+                                //Delete file
+                                unlink($filePathImageSize);
+
+                                //Unset
+                                unset($filePathImageSize);
+
+
+                            }
+
+
+                        }
+
+                        //delete original image
+                        unlink($fileOriginalPath);
+
+                        //Delete photo
+                        $photo->delete();
+
+                        unset($photo_name);
+
+
+
+                    }
+
+                    //Delete foresale
+                    $page->delete();
+
+                    $response = array( 'status'=> 'success', 'message'=> 'successfully deleted.' );
+
+                }else{
+                    $response = array( 'status'=> 'fail', 'message'=> 'Unauthorized request.' );
+                }
+
+
+
+            }else{
+                $response = array( 'status'=> 'fail', 'message'=> 'Unauthorized request.' );
+            }
+
+        }
+
+        return $response ;
     }
+
+
